SBERDOMA-378 Created default render for text
