+++ apps/condo/domains/organization/hooks/useTableColumns.tsx
@@ -10,7 +10,7 @@ import { OrganizationEmployeeRole } from '../utils/clientSchema'
 import { getTextFilterDropdown, getFilterIcon, FilterContainer } from '@condo/domains/common/components/TableFilter'
 import { isEmpty } from 'lodash'
 import { Highliter } from '../../common/components/Highliter'
-import { EmptyTableCell } from '@condo/domains/common/components/EmptyTableCell'
+import { EmptyTableCell } from '@condo/domains/common/components/Table/EmptyTableCell'
 
 const getFilteredValue = (filters: IFilters, key: string | Array<string>): FilterValue => get(filters, key, null)
